requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
pyarrow
Pillow
matplotlib
plotly
python-dotenv